# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BOND-GYARADOS-API",
    "description": "En esta documentacion encontrara toda la informacion relacionada con la adquicion de informacion para el proyecto bond-gyarados-api y su ulizacion.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://<NameResource>.azurewebsites.net/api/<NameFunction>/<NameOperation>"
    }
  ],
...
  "paths": {
    "/api/users/add": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Operation to add users",
        "description": "Add new users",
        "operationId": "postAddUser",
        "requestBody": {
          "description": "Operation about users",
          "content": {
            "add User": {
              "schema": {
                "$ref": "#/components/schemas/UserAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseAdd"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/users/login": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Operation to get status users",
        "description": "Get status users",
        "operationId": "getLoginUser",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "Value for user",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "operator"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "Value for password",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": 123456
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLogin"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/users/update": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Operation to update users",
        "description": "Update users",
        "operationId": "postUpdateUser",
        "requestBody": {
          "description": "Operation about users",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "Response Update": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUpdate"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/users/delete": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Operation to delete users",
        "description": "Delete users",
        "operationId": "postDeleteUser",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "Value for user",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "operator"
            }
          },
          {
            "name": "confirmation",
            "in": "query",
            "description": "Confirmation delete",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDelete"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserAdd": {
        "required": [
          "user",
          "email",
          "password",
          "state",
          "role"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "operator"
          },
          "email": {
            "type": "string",
            "example": "operator@enterprise.com"
          },
          "password": {
            "type": "string",
            "example": 123456
          },
          "state": {
            "type": "string",
            "example": "activate"
          },
          "role": {
            "type": "string",
            "example": "admin"
          }
        }
      },
      "UserUpdate": {
        "required": [
          "user",
          "email",
          "password",
          "state",
          "role"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "operator"
          },
          "email": {
            "type": "string",
            "example": "operator@enterprise.com"
          },
          "password": {
            "type": "string",
            "example": 123456
          },
          "state": {
            "type": "string",
            "example": "activate"
          },
          "role": {
            "type": "string",
            "example": "admin"
          }
        }
      },
...
      "ResponseLogin": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "operador"
          },
          "state": {
            "type": "string",
            "example": "activate"
          },
          "role": {
            "type": "string",
            "example": "admin"
          },
          "access": {
            "type": "string",
            "example": "permited"
          }
        }
      },
      "ResponseAdd": {
        "required": [
          "user",
          "email",
          "password",
          "state",
          "role"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "operator"
          },
          "email": {
            "type": "string",
            "example": "operator@enterprise.com"
          },
          "password": {
            "type": "string",
            "example": 123456
          },
          "state": {
            "type": "string",
            "example": "activate"
          },
          "role": {
            "type": "string",
            "example": "admin"
          },
          "statusUser": {
            "type": "string",
            "example": "added"
          }
        }
      },
      "ResponseUpdate": {
        "required": [
          "user",
          "email",
          "password",
          "state",
          "role"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "operator"
          },
          "email": {
            "type": "string",
            "example": "operator@enterprise.com"
          },
          "password": {
            "type": "string",
            "example": 123456
          },
          "state": {
            "type": "string",
            "example": "activate"
          },
          "role": {
            "type": "string",
            "example": "admin"
          },
          "statusUser": {
            "type": "string",
            "example": "update"
          }
        }
      },
      "ResponseDelete": {
        "required": [
          "user",
          "email",
          "password",
          "state",
          "role"
        ],
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "operator"
          },
          "statusUser": {
            "type": "string",
            "example": "delete"
          }
        }
      }
    }
  }
}